Update ui structure and create organizer-ui
