+++ easyticket/.idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.13 (easyticket) (2)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.13 (easyticket)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.13 (easyticket) (3)" project-jdk-type="Python SDK" />
 </project>
