+++ easyticket/app/templates/layout/header.html
@@ -1,21 +1,41 @@
 <div>
     <header class="gradient-bg text-white shadow-lg">
-        <div class="container mx-auto px-4 py-6">
-            <div class="flex items-center justify-between">
-                <h1 class="text-3xl font-bold">🎉 EasyTicket</h1>
-                <nav class="hidden md:flex space-x-6">
-                    <a href="{{url_for('index')}}" class="hover:text-yellow-300 transition-colors">Trang chủ</a>
-                    <a href="#" class="hover:text-yellow-300 transition-colors">Sự kiện</a>
-                    <a href="#" class="hover:text-yellow-300 transition-colors">Liên hệ</a>
-                     {% if current_user.is_authenticated %}
-                    <a href="#" class="hover:text-yellow-300 transition-colors">Xin chào, {{ current_user.username }}</a>
-                    <a href="{{ url_for('auth.logout') }}" class="hover:text-yellow-300 transition-colors">Đăng xuất</a>
-                {% else %}
-                    <a href="{{ url_for('auth.login') }}" class="hover:text-yellow-300 transition-colors">Đăng nhập</a>
-                    <a href="{{ url_for('auth.register') }}" class="hover:text-yellow-300 transition-colors">Đăng ký</a>
-                {% endif %}
+        <div class="container py-3">
+            <div class="d-flex align-items-center justify-content-between">
+                <h1 class="h3 mb-0">
+                    <i class="fas fa-ticket-alt me-2"></i>EasyTicket
+                </h1>
+                <nav class="d-none d-md-flex">
+                    <a href="{{ url_for('index') }}" class="text-white text-decoration-none me-4 hover-underline">
+                        <i class="fas fa-home me-1"></i>Trang chủ
+                    </a>
+                    <a href="#" class="text-white text-decoration-none me-4 hover-underline">
+                        <i class="fas fa-calendar me-1"></i>Sự kiện
+                    </a>
+                    <a href="#" class="text-white text-decoration-none me-4 hover-underline">
+                        <i class="fas fa-phone me-1"></i>Liên hệ
+                    </a>
+                      {% if current_user.is_authenticated %}
+                    <a href="#" class="text-white text-decoration-none me-4 hover-underline">
+                        Xin chào, {{ current_user.username }}
+                    </a>
+                    <a href="{{ url_for('auth.logout') }}" class="text-white text-decoration-none me-4 hover-underline">
+                        <i class="fas fa-user me-1"></i>Đăng xuất
+                    </a>
+                        {% else %}
+                    <a href="{{ url_for('auth.login') }}" class="text-white text-decoration-none me-4 hover-underline">
+                        <i class="fas fa-user me-1"></i>Đăng nhập
+                    </a>
+                       <a href="{{ url_for('auth.register') }}" class="text-white text-decoration-none me-4 hover-underline">
+                        <i class="fas fa-user me-1"></i>Đăng ký
+                    </a>
+                    {% endif %}
                 </nav>
+                <button class="navbar-toggler d-md-none text-white" type="button">
+                    <i class="fas fa-bars"></i>
+                </button>
             </div>
         </div>
     </header>
+
 </div>
